refactor: move api run function to different package for testing
