+++ .github/workflows/job_test_go_api_local.yaml
@@ -20,3 +20,5 @@ jobs:
       - name: Test
         run: task test
         working-directory: go
+        env:
+          SIMULATON_TEST: true
