+++ go/Taskfile.yml
@@ -8,6 +8,8 @@ tasks:
       - go fmt ./...
       - task: lint
   test:
+    env:
+      SIMULATON_TEST: true
     cmds:
       - docker pull mysql:latest
       - go test -cover -json -failfast ./... | tparse -all -progress  -smallscreen
@@ -24,7 +26,3 @@ tasks:
   lint:
     cmds:
       - golangci-lint run
-
-  simulate:
-    cmds:
-      - go run github.com/jellevandenhooff/gosim/cmd/gosim test -v -run TestGosim
