+++ go/apps/api/routes/v2_ratelimit_delete_override/200_test.go
@@ -8,13 +8,13 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-	handler "github.com/unkeyed/unkey/go/cmd/api/routes/v2_ratelimit_get_override"
+	handler "github.com/unkeyed/unkey/go/apps/api/routes/v2_ratelimit_delete_override"
 	"github.com/unkeyed/unkey/go/pkg/db"
 	"github.com/unkeyed/unkey/go/pkg/testutil"
 	"github.com/unkeyed/unkey/go/pkg/uid"
 )
 
-func TestGetOverrideSuccessfully(t *testing.T) {
+func TestDeleteOverrideSuccessfully(t *testing.T) {
 	ctx := context.Background()
 	h := testutil.NewHarness(t)
 
@@ -31,17 +31,14 @@ func TestGetOverrideSuccessfully(t *testing.T) {
 
 	// Create an override
 	identifier := "test_identifier"
-	limit := int32(10)
-	duration := int32(1000)
 	overrideID := uid.New(uid.RatelimitOverridePrefix)
-
 	err = db.Query.InsertRatelimitOverride(ctx, h.DB.RW(), db.InsertRatelimitOverrideParams{
 		ID:          overrideID,
 		WorkspaceID: h.Resources.UserWorkspace.ID,
 		NamespaceID: namespaceID,
 		Identifier:  identifier,
-		Limit:       limit,
-		Duration:    duration,
+		Limit:       10,
+		Duration:    1000,
 		CreatedAt:   time.Now().UnixMilli(),
 	})
 	require.NoError(t, err)
@@ -55,44 +52,30 @@ func TestGetOverrideSuccessfully(t *testing.T) {
 
 	h.Register(route)
 
-	rootKey := h.CreateRootKey(h.Resources.UserWorkspace.ID, fmt.Sprintf("ratelimit.%s.read_override", namespaceID))
+	rootKey := h.CreateRootKey(h.Resources.UserWorkspace.ID, fmt.Sprintf("ratelimit.%s.delete_override", namespaceID))
 
 	headers := http.Header{
 		"Content-Type":  {"application/json"},
 		"Authorization": {fmt.Sprintf("Bearer %s", rootKey)},
 	}
 
-	// Test getting by namespace name
-	t.Run("get by namespace name", func(t *testing.T) {
+	// Test deleting by namespace name
+	t.Run("delete by namespace name", func(t *testing.T) {
 		req := handler.Request{
 			NamespaceName: &namespaceName,
 			Identifier:    identifier,
 		}
 
 		res := testutil.CallRoute[handler.Request, handler.Response](h, route, headers, req)
-		require.Equal(t, 200, res.Status, "expected 200, received: %v", res.Body)
-		require.NotNil(t, res.Body)
-		require.Equal(t, overrideID, res.Body.OverrideId)
-		require.Equal(t, namespaceID, res.Body.NamespaceId)
-		require.Equal(t, identifier, res.Body.Identifier)
-		require.Equal(t, int64(limit), res.Body.Limit)
-		require.Equal(t, int64(duration), res.Body.Duration)
-	})
+		require.Equal(t, 200, res.Status, "expected 200, received: %s", res.RawBody)
 
-	// Test getting by namespace ID
-	t.Run("get by namespace ID", func(t *testing.T) {
-		req := handler.Request{
-			NamespaceId: &namespaceID,
-			Identifier:  identifier,
-		}
+		// Verify the override was deleted (check soft delete)
+		override, err := db.Query.FindRatelimitOverrideById(ctx, h.DB.RO(), db.FindRatelimitOverrideByIdParams{
+			WorkspaceID: h.Resources.UserWorkspace.ID,
+			OverrideID:  overrideID,
+		})
 
-		res := testutil.CallRoute[handler.Request, handler.Response](h, route, headers, req)
-		require.Equal(t, 200, res.Status, "expected 200, received: %v", res.Body)
-		require.NotNil(t, res.Body)
-		require.Equal(t, overrideID, res.Body.OverrideId)
-		require.Equal(t, namespaceID, res.Body.NamespaceId)
-		require.Equal(t, identifier, res.Body.Identifier)
-		require.Equal(t, int64(limit), res.Body.Limit)
-		require.Equal(t, int64(duration), res.Body.Duration)
+		require.NoError(t, err)
+		require.True(t, override.DeletedAtM.Valid, "Override should be marked as deleted")
 	})
 }
