+++ go/apps/api/routes/v2_ratelimit_delete_override/401_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	handler "github.com/unkeyed/unkey/go/cmd/api/routes/v2_ratelimit_delete_override"
+	handler "github.com/unkeyed/unkey/go/apps/api/routes/v2_ratelimit_delete_override"
 	"github.com/unkeyed/unkey/go/pkg/testutil"
 )
